+++ src/main/java/com/delgo/api/comm/quartz/job/RefreshPriceJob.java
@@ -1,8 +1,8 @@
 package com.delgo.api.comm.quartz.job;
 
 import com.delgo.api.domain.room.Room;
-import com.delgo.api.service.PriceService;
 import com.delgo.api.service.RoomService;
+import com.delgo.api.service.crawling.GetPriceCrawlingService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.quartz.JobExecutionContext;
@@ -18,7 +18,7 @@ import java.util.List;
 public class RefreshPriceJob extends QuartzJobBean {
 
     private final RoomService roomService;
-    private final PriceService priceService;
+    private final GetPriceCrawlingService getPriceCrawlingService;
 
     @Override
     protected void executeInternal(JobExecutionContext jobExecutionContext) throws JobExecutionException {
@@ -27,11 +27,10 @@ public class RefreshPriceJob extends QuartzJobBean {
 
         // 스케줄러에 의해 돌아갈 코드 작성
         List<Room> roomList = roomService.selectAll();
-        priceService.crawlingProcess(roomList);
+        getPriceCrawlingService.crawlingProcess(roomList);
 
         LocalDateTime endTime = LocalDateTime.now();
         System.out.println("************ RefreshPriceJob Exit ************" + endTime);
         System.out.println("총 걸린 시간 : " + ChronoUnit.MINUTES.between(startTime, endTime));
-
     }
 }
